Fixed strange Unity issue
Had an issue with Unity refusing to open the custom editor windows. Fixed it by creating new EditorWindow scripts and copy pasting the content of the old window scripts into them. The old scripts were deleted

Also added  a log message that gets displayed in the unity console with information on where to find the ShapesWindow tool